ops: remove circular deps & add restore
