+++ apps/backend-ops/src/firebaseSetup.ts
@@ -3,69 +3,11 @@
  *
  * This module provides functions to trigger a firestore restore and test user creations.
  */
-import request from 'request';
 import { appUrl } from '@env';
 import { syncUsers } from './users';
 import { upgradeAlgoliaMovies, upgradeAlgoliaOrgs } from './algolia';
 import { migrate } from './migrations';
-
-function getRestoreURL(appURL: string): string {
-  return `${appURL}/admin/data/restore`;
-}
-
-function getBackupURL(appURL: string): string {
-  return `${appURL}/admin/data/backup`;
-}
-
-/**
- * Trigger a firestore database restore operation for the given project
- */
-export async function restore(appURL: string) {
-  if (process.env['ADMIN_PASSWORD'] === undefined) {
-    throw new Error('no ADMIN_PASSWORD in your env, we need this to trigger backups / restores');
-  }
-
-  const url = getRestoreURL(appURL);
-  const form = {
-    password: process.env['ADMIN_PASSWORD']
-  };
-
-  // promisified request
-  return new Promise((resolve, reject) => {
-    request.post({ url, form }, (error, response) => {
-      if (error) {
-        reject(error);
-      } else {
-        resolve(response);
-      }
-    });
-  });
-}
-
-/**
- * Trigger a firestore database backup operation for the given project
- */
-export async function backup(appURL: string) {
-  if (process.env['ADMIN_PASSWORD'] === undefined) {
-    throw new Error('no ADMIN_PASSWORD in your env, we need this to trigger backups / restores');
-  }
-
-  const url = getBackupURL(appURL);
-  const form = {
-    password: process.env['ADMIN_PASSWORD']
-  };
-
-  // promisified request
-  return new Promise((resolve, reject) => {
-    request.post({ url, form }, (error, response) => {
-      if (error) {
-        reject(error);
-      } else {
-        resolve(response);
-      }
-    });
-  });
-}
+import { restore } from './admin';
 
 export async function prepareForTesting() {
   console.info('Syncing users...');
